requirements.txt
requests
pandas
numpy
python-dotenv
pyyaml
schedule
tenacity
pytest
requests-mock